:recycle: refactor: 코드리뷰 반영
- 가게로 패키지 이동
