+++ src/main/java/com/sparta/oishitable/domain/owner/restaurant/menus/service/MenuService.java
@@ -1,17 +1,16 @@
-package com.sparta.oishitable.domain.owner.menus.service;
+package com.sparta.oishitable.domain.owner.restaurant.menus.service;
 
 import com.sparta.oishitable.domain.auth.service.AuthService;
-import com.sparta.oishitable.domain.owner.menus.dto.request.MenuCreateRequest;
-import com.sparta.oishitable.domain.owner.menus.dto.request.MenuUpdateRequest;
-import com.sparta.oishitable.domain.owner.menus.dto.response.MenuDetailResponse;
-import com.sparta.oishitable.domain.owner.menus.dto.response.MenuFindResponse;
-import com.sparta.oishitable.domain.owner.menus.entity.Menu;
-import com.sparta.oishitable.domain.owner.menus.repository.MenuRepository;
+import com.sparta.oishitable.domain.owner.restaurant.menus.dto.request.MenuCreateRequest;
+import com.sparta.oishitable.domain.owner.restaurant.menus.dto.request.MenuUpdateRequest;
+import com.sparta.oishitable.domain.owner.restaurant.menus.dto.response.MenuDetailResponse;
+import com.sparta.oishitable.domain.owner.restaurant.menus.dto.response.MenuFindResponse;
+import com.sparta.oishitable.domain.owner.restaurant.menus.entity.Menu;
+import com.sparta.oishitable.domain.owner.restaurant.menus.repository.MenuRepository;
 import com.sparta.oishitable.domain.owner.restaurant.entity.Restaurant;
 import com.sparta.oishitable.domain.owner.restaurant.service.OwnerRestaurantService;
 import com.sparta.oishitable.global.exception.NotFoundException;
 import com.sparta.oishitable.global.exception.error.ErrorCode;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,10 +26,10 @@ public class MenuService {
     private final MenuRepository menuRepository;
 
     @Transactional
-    public void createMenu(Long userId, Long restaurantId, List<@Valid MenuCreateRequest> request) {
+    public void createMenu(Long userId, Long restaurantId, List<MenuCreateRequest> request) {
         Restaurant restaurant = restaurantService.findById(restaurantId);
 
-        authService.checkAuthenticationUser(restaurant.getOwner().getId(), userId);
+        authService.checkUserAuthority(restaurant.getOwner().getId(), userId);
 
         List<Menu> menus = request.stream()
                 .map(m -> m.toEntity(restaurant))
@@ -57,7 +56,7 @@ public class MenuService {
     public void updateMenu(Long userId, Long restaurantId, Long menuId, MenuUpdateRequest request) {
         Menu menu = findMenuByRestaurantIdAndId(restaurantId, menuId);
 
-        authService.checkAuthenticationUser(menu.getRestaurant().getOwner().getId(), userId);
+        authService.checkUserAuthority(menu.getRestaurant().getOwner().getId(), userId);
 
         menu.update(request.menuName(), menu.getPrice(), menu.getDescription());
     }
@@ -66,7 +65,7 @@ public class MenuService {
     public void deleteMenu(Long userId, Long restaurantId, Long menuId) {
         Menu menu = findMenuByRestaurantIdAndId(restaurantId, menuId);
 
-        authService.checkAuthenticationUser(menu.getRestaurant().getOwner().getId(), userId);
+        authService.checkUserAuthority(menu.getRestaurant().getOwner().getId(), userId);
 
         menuRepository.delete(menu);
     }
